AIDEGen: Load default dependent jars from module-info.json.
Some core modules like core-or and core-libart are load as default
dependencies by build system so that the developers no need to write the
dependent modules in Android.mk or Android.bp. By aosp/875753, it can
collects the dependent jar files in module-info.json and then the AIDEGen
can trace all dependencies of target project.

Bug: 122557164
Test: 1. m aidegen
      2. $ANDROID_HOST_OUT/bin/aidegen DocumentsUI
      3. Open packages/apps/DocumentsUI/src/com/android/documentsui/
         services/CopyJob.java
      4. Check import android.system and import libcore.io.IoUtils not
         in red.

Change-Id: I4fdb987639cc1b26fed544562dbf0eb1bdda64a1