AIDEGen: Use the parent folder of ./idea and .iml as argument to launch IDEA.
IntelliJ doesn't treat .iml file as a project file, but a module description
file, so the patch use the parent folder of .idea instead of the path of .iml
to launch IntelliJ. To IntelliJ, there're only 2 kinds of project, one is file
base, the .irp file, and the other is directory base, which contains .idea
folder and .iml file(s). So, launch IntelliJ with a .iml file is same meaning
as to let it open a text file. IntelliJ needs the correct project content path
for opening a project.

Bug: 124460293
Test: Open a workable IntelliJ project by IDEA. And make sure the IDEA version
      you used for the test.
      1. $ m aidegen && $ANDROID_HOST_OUT/bin/aidegen-dev cts -r
      2. In the option list, pick up the same IDEA version as the opened one
      3. make sure the other IDEA instance will launch and open the cts project
         that AIDEGen created
      4. $ atest aidegen_unittests, and it should be all pass

Change-Id: I0ddc8c9f6886e4944a414702617611c8000fd238