AIDEGen: Unnable to create correct project file for frameworks/native
We originally support Java files and its related libraries only, but now
we should deal with C, C++ files and their related libraries as well. In
this way, the project view will show the whole frameworks/native but not
frameworks/native/opengl/tests only, which contains java files in it.

Bug: 126472374
Test: 1. m aidegen
      2. $ANDROID_HOST_OUT/bin/aidegen-dev frameworks/native -v
         The folders should not be only frameworks/native/opengl/tests.

Change-Id: I90d5714e1e9427457d4f8dafd7442ccd809fa5e2