AIDEGen: Build the aapt2.srcjar to generate R.java.
Originally aidegen builds APPS modules if the R.java does not exist. But it
will causes build break by system.img oversized.
This CL will build the aapt2.srcjar of module if the R.java does not exist.
And it will not create the module's apk so there is no system oversized
issue from aidegen.

Bug: 128512508
Test: 1. Checkout internal master and lunch walleye-eng
      2. rm -rf out/target/common/obj/APPS/Settings_intermediates
      3. m installclean
      4. m aidegen;aidegen-dev Settings framework cts vendor/xts
      5. m -j should not failed.
      6. Check out/target/common/obj/APPS/Settings_intermediates/aapt2
         exists.
      7. Double check: go through step 2~6 but replace step 3 to m clean

Change-Id: If04853b86a6e63d3f79f4e03c4c9ba1dbdef1ebb