AIDEGen: Don't assemble jar path when module already has "installed" path
In _set_jars_jarfile it assembles the module path and jar path to get jar path.
However, it may generate wrong path to make build fail.
If the module has 'installed' path in its module_info data, we should make use
of it. This CL changes design to use the 'installed' path first.

Bug: 132301301
Test: 1. m clean -j16
      2. m aidegen; aidegen-dev framework
      3. you should see the log atest_utils.py:154:INFO: Build successful

Change-Id: Ie5bcaee941ce221a179fed18ac198261a85e054e