AIDEGen: Prevent generating module_bp_java_deps.json error.
Currently we use chdir() to change the directory to the AOSP root folder
and then AIDEGen could generate the iml files for the whole AOSP tree.
But the action of changing current work directory will cause the
generating of module_bp_java_deps.json failed in some corner case. In
order to prevent from incorrect using of chdir(), we can set the
absolute and relative paths of the whole AOSP project to correct it.

Bug: 139264620
Bug: 136623722
Test: 1. m aidegen
      2. aidegen-dev -a framework
      3. If your AOSP root folder is ~/aosp/, check the aosp.iml
         and the dependencies.iml exist in ~/aosp/
      4. rm out/soong/module_bp_java_deps.json
      5. cd packages/apps/Settings
      6. aidegen-dev -a
      7. Confirm the out/soong/module_bp_java_deps.json is generated.
      8. Check if the whole Android tree is opened in IntelliJ.

Change-Id: I3807a33c61f51ba7c4321ade845a331ec1625115