AIDEGen: Remove DISABLE_ROBO_RUN_TESTS environment variable
AIDEGen ran command 'mma frameworks/base' to build the module for
generating the related jar or R.java. But it triggered the issue
running the robolectric test of the module. Therefore, we added the
environment variable DISABLE_ROBO_RUN_TESTS in run_robotests.mk for
AIDEGen to skip the robolectric build. But now AIDEGen only generates
*.jar or *.srcjar by command 'm *.jar *.srcjar'. So, we can remove the
environment variable now.

Bug: 147409906
Test: 1. m clean
      2. m aidegen
      3. aidegen-dev framework
      4. m aidegen
      5. Confirm the string "Environment variable DISABLE_ROBO_RUN_TESTS
         was modified (true => ), regenerating..." is not shown.

Change-Id: I8f405c29ad4dabd6802c384668a2510dbda5574e