AIDEGen: fix aidegen -i s native_module fail case
When used aidegen -i s to launch Android Studio with native module,
the file IO relevant error, e.g., NotADirectoryError happened. The
ide_util_obj of Android Studio can’t judge if it’s a launch action
without configuring the java IDE so the native case ran into Java
IDE configuration and caused the error. Because the java project is
a .idea directory and the native is a file, using the difference as
a temporary solution to prevent the messed up.

Bug: 150273132
Test: 1. m aidegen;aidegen-dev -i s frameworks/compile/libbcc/bcinfo/BitReader_3_0
         2. Android Studio launched and open with BitReader_3_0
Change-Id: I59825cad2285c10417a3cbe4fa370e57facc569e