AIDEGen: provide choice of the preferred language
Right now, the behavior to launch both CLion and IntelliJ IDEA causes a
bug. When they are both launched to open their projects which aidegen
generated under the directory frameworks, sometimes the race condition
happened. In my case, the IDEA failed to load the content source of its
main module. And it can recover when I restart the same IDEA project.
Since it's rare that users need to use 2 IDEs at the same time, we could
provide an option to them to get the preferred one and launch it only.
It also solves the case that our exisiting java user to see the other
annoying IDE launched and then they need to close it by themselves.

Bug: 150578306

Test: 1. m aidegen;aidegen-dev frameworks/native
      2. Input 1 or 2 to select the IDE you want to invoked.
      3. The relevant IDE should launch correctly.
Change-Id: Ia68eb1f95e2c1050196271885e8d408d16ca85bb