AIDEGen: Keeps the order of source files from the module-info.json.
Since the order of srcjar files sequence impacts the dependencies,
keeping them in the original order on the module-info.json is important.
This CL keeps the order of the source_locator.ModuleData into
project_info.ProjectInfo. In single project case, such as framework,
the content of some iml files is generated from the
source_locator.ModuleData. It will be fine. In multiple projects'
case, the ProjectInfo class merges from modules' data of different
projects, If it doesn't keep the order of ModuleData, it might cause
some duplicated class issue.

Bug: 153694169
Test: 1. sh run_tests.sh
      2. atest aidegen_unittests --host
      3. Increase the code voverage of source_locator.py to 92%.
Change-Id: I7e123f7a79d1d8805eed15151390540b84c33cc1