AIDEGen: Fix potential issues of str.startswith judgement.
There are potential issues when we determine if the path is under
another path using 'path.startswith(another_path)' judgement. For
example, if the path is "system/security/keystore-engine" which is
certainly not under the path "system/security/keystore" but the
return of 'startswith' judgement is True. To fix it, I replace
'startswith' with 'is_source_under_relative_path(path, another_path)'
function to determine if a path is under another path.

Bug: 155585344
Test: 1. sh run_tests.sh
         All unittests pass!
Change-Id: Icd1cfd0f0b88ff07d87591deb00ddf3d219b63ef