Atest: install/update coverage while running unit tests.
Bug: 159094042

1. run_atest_unittests.sh coverage
If coverage is not installed, it will be install before running tests.
if the interpreter of coverage is /usr/bin/python, it will be replaced
with /usr/bin/python3 and move forward to testing.

2. atest_run_unittests.py coverage
If coverage is not installed, it will only run unit tests only because
the infra will not be allowed to install anything.

Test: source build/envsetup.sh; lunch
  ./run_atest_unittests.sh coverage # ~10s
  ./atest_run_unittests.py coverage --show-missing # ~3s
  the above commands behave the same and should run successfully without error.

  ./atest_run_unittests.py coverage
  Can only display code coverage.

Change-Id: I47eee9494269e24ac843c0ed2ca92ddc3cd4f104