Use the correct helper script path location
The atest_tradefed.sh script assumes that the helper script it uses is in the same directory. This is not the case when run from Bazel.

This change adds an environment variable to allow customizing the atest helper script location. This variable is used to generate a Bazel test executable script with the helper script path set to the correct location in the runfiles tree.

Bug: 200988722
Test: ./run_bazel_mode_atest_unittest.sh
Change-Id: Ief591ce07712cde3b59ebd03f300bd36661d8d28