Write the module_name attribute for prebuilt targets
The Bazel test rule generates a test executable script with an empty module name without this attribute. This causes Tradefed to fail since it requires the module name on the command line.

Note that this attribute will also be eventually used by Atest to query the Soong modules to pre-build.

Bug: 200988722
Bug: 204771853
Test: atest --host atest_unittests
Change-Id: Iba560b9fdb38e3484c6912a85416b617779512ae