Push dependency information into target classes
Add a method to Target that enables the different implementations to
declare their dependencies. These references are later resolved by the
generator into actual Target objects.

This approach pushes information out of the generator and consolidates
it in the Target sub-classes. An additional benefit is that this makes
it possible to explicitly resolve dependencies using an iterative DFS
implementation which is more Pythonic.

The unit tests are also refactored to depend less on implementation
details and remove irrelevant information out of the test body for
readability.

Test: atest --host atest_unittests
Bug: 192581397
Change-Id: I547793e83dccfc2aed8627492566a368b64601da