Use the suite to determine if a test is deviceless
Bazel mode currently checks whether a test requires a device by relying
on the 'is_unit_test' field in module-info. This approach is not
guaranteed to correctly run tests without a connected device. Instead
we check that the test is in the 'host-unit-tests' compatibility suite
which is aligned with what the CI runs on the deviceless cluster.

Bug: 205918165
Test: atest --host atest_unittests
Change-Id: I360478a47eed36399c7abecb16ddda209a03c2fc