Ensure that soong_prebuilt files are never empty
Ensure that soong_prebuilt targets always have at least one file to
avoid evaluation errors when running Bazel cquery on a clean tree to
find dependencies.

This happens because soong_prebuilt dependency target globs don't match
any files when the workspace symlinks are broken and point to build
artifacts that still don't exist. This in turn causes errors in rules
that reference these targets via attributes with allow_single_file=True
and which expect a file to be present.

This change adds a placeholder file when the glob doesn't match any
paths to ensure that rule implementations that require at least one item
can be properly evaluated in cquery invocations.

Test: m clean && atest-src --bazel-mode --dry-run && pushd ${ANDROID_BUILD_TOP}/out/atest_bazel_workspace && bazel cquery "deps(//tools/asuite/atest:atest_unittests_host)"
Bug: 206843290
Change-Id: Ie871d013cd8dc2211e9926d80e33169b3f703112