Remove unnecessary conditions in runtime_deps
Modify the generator to no longer include a condition in the
runtime_deps select expression for configurations that are not supported
by the target. For example, a host-only target no longer includes a
device condition even if the dependency supports it.

Note that while dependencies with no compatible configuration are no
longer included in a target's runtime_deps attribute list, the
dependency target is still generated which could potentially leave it
unreferenced.

Bug: 206896017
Test: atest --host atest_unittests
Test: atest-src --bazel-mode --host atest_unittests
Change-Id: If47f01f5701d90db54a5bca9b67bfad9f777c668