Add integration tests for Bazel mode
Add basic integration tests for the Atest Bazel mode feature. The tests
run Atest on a source tree to verify end-to-end behavior that is not
unit testable.

The tests can be run using Atest as indicated in this change's commit
message Test stanza.

Note that these tests are not added to Test Mapping since they require
source in their current form and can't run on the standard CI
infrastructure.

Bug: 215483376
Test: atest --host bazel_mode_test
Change-Id: Ia88fd3d351f658f7ba4551589497e0ea4161c975