Pass Android JDK Path to Bazel for workspace generation
Pass the Android JDK path to Bazel when calling cquery during Atest
workspace generation. This fixes the Java issues that were seen in CI.

This change also changes the Bazel `output_user_root` to a unique temporary directory instead of a hard-coded location in /tmp. This avoids issues with reading stale files from a previous run.

Test: source build/envsetup.sh; lunch aosp_cf_x86_64_phone-userdebug; gen_workspace_archive_test.sh
Bug: 202790559
Change-Id: I0e79bd80365a674ac13b1ec7977eb1fa0e9a6698