Default to running the device variant when Bazel mode enabled
Fix an issue in Bazel mode where the host variant of a test is
incorrectly run instead of the device variant. The behavior should
mimic standard mode and should default to running the device variant
unless the  --host  command-line argument is specified. Tests with only
a host variant can skip the flag.

Bug: 226389735
Test: atest --host atest_unittests
Change-Id: Iaad38d3615238ce516f5fc16c0df5f8dbe79980d