Install target app before running an instrumentation test
This change enables TradefedRunner to assemble `--module-arg` argument
to install the target apk found in previous changes. As a result
Tradefed invokes `TestAppInstallSetup` to handle the installation
process.

With verbose mode on the users are able to see path of Settings.apk and
Contacts.apk are assembled with --module-arg in the Tradefed command.

Bug: 177626045

Test: atest-src -v --dry-run SettingsUnitTests ContactsTests
Test: atest-src -v  SettingsUnitTests ContactsTests

Change-Id: I9793201a0071332538141e431f8b291c49fabf7f