Add a Tradefed runner for Atest Bazel mode tests
This runner enables running tests in the workspace generated by Atest Bazel mode on TradeFed-managed hosts. The implementation expects an archive containing the workspace that it downloads and extracts on the host. Tests are then run with Bazel using the configured options.

See `atest/bazel/scripts/gen_workspace_archive.sh` for how to generate the workspace archive.

Test: atest bazel-test-runner-tests
Bug: 202790559
Change-Id: I2691199499b3d5f6470f4f0b9769ef317572deb0