Avoid KeyError when running native tests
Atest is not able to parse header files yet, so it's possible to parse
testcases and test methods incompletely. This CL is just a small fix to
avoid KeyError only: only set up values when the key exists(found
class name); as the result, Atest will exit with test not found instead
of KeyError.

BUG: 234531695

Test: (internal) atest-src LwisClientTest --no-bazel-mode
Change-Id: Ie0ea88b3c83bd7b66339a8f322a523ed4b9340d3