Probe target apks of instrumentation tests
Based on previous change list, Atest is able to probe target apk and
exclude persistent one. This change appends the target module to the
build_targets so that the target apk can be generated in the build time;
and stores the absolute path of the target app in test_info, so in the
next CL the test runner would be able to assemble the Tradefed command
to installed the found apk.

With verbose mode on, users are able to see informative messages like:
"Found essentail artifacts: [..]" and
"Settings" and "Camera2" were added to the build_targets.

Bug: 177626045

Test: atest SettingsUnitTests -v --dry-run
Test: atest-src SettingsUnitTests Camera2Tests CtsTimeTestCases FrameworksCorePackageManagerTests -v --dry-run

Change-Id: Id890bdf3cb48a5e7deaaa285f794faee745c3137