Change the TF test dir to the base of the target's runfiles tree
TradeFed cannot find some files because its test directory is set to a subdirectory in the runfiles tree. For example, `art-run-test-004-checker-UnsafeTest18` declares `art-run-test-checker` in the `host_required` attribute which falls outside this subdirectory. Since the files cannot be found at runtime, the tests fail.

This change sets TF's test directory to the root of the target's runfiles tree and widens the scope of TF's search.

Test: atest-src atest_unittests
Test: atest-src art-run-test-004-checker-UnsafeTest18 \
          --experimental-device-driven-test
Test: atest-src -c -m --host \
          --host-unit-test-only \
          --bazel-arg=--test_timeout=3000
Bug: 222437352
Change-Id: I8f74a5420f2eb301c9735cfb9fa18b6dfc756e18