DONT index when plocate.db remains the same
BUG: 206886222

Previously given that the src does not change, the data checksum of
plocate.db may still be different for the metadata changes. In the
experiment the same src may result in different checksum of the db, the
file size will reamin the same. This CL determines whether Atest needs
to run index by telling both md5sum and the filesize; if the filesize
remains the same, do not run index as well.

Test: atest-src hello_world_test -v
Change-Id: I4edc94b3bd1c3a7dd9be69d95d4babcee2b26cda