Run generate_buildfiles_checksum() in parallel
`generate_buildfiles_checksum` calculates the checksum of every
Android.bp and Android.mk for determining whether Atest needs to rebuild
module-info.json.

Measurements show that the method takes 0.88s to calculate the checksums
of 9810 files and stores the result in the disk, and 3.00s of 23208
files. Running `generate_buildfiles_checksum` in parallel can save the
whole method invocation time.

Test: atest hello_world_test --host --test
Test: atest-src hello_world_test --host --test

Bug: 229761444

Change-Id: Ide334dc779bbefb822beb1515f64e730c029191e