Add `vts-core-tradefed-harness` for VTS tests
Currently we add `compatibility-tradefed` for VTS tests by default in the `tradefed_device_test` macro. But besides `compatibility-tradefed`, some VTS tests also require other harness dependencies that are included in the `vts-core-tradefed-harness`. Moreover, the `vts-core-tradefed-harness` includes `compatibility-tradefed`. Hence, this change adds the `vts-core-tradefed-harness` for VTS tests instead of the `compatibility-tradefed` in the `tradefed_device_test` macro.

Bug: 220991354
Test: atest-src atest_unittests
Test: atest-src vts_treble_vintf_framework_test --experimental-device-driven-test --experimental-java-runtime-dependencies
Change-Id: I1f1d5dbebef023aef6aea90a6fde9a27d9444c3d