Create a dedicated directory for Tradefed test files
We currently specify the base of the Bazel test target's runfiles tree
as the TF test directory. This causes problems because the layout TF
sees is different than what it usually expects. It also diverges from
CI behavior and causes surprises. For example, a test could pass
locally because TF finds a file it needs but could fail in presubmit.

Instead, we create a dedicated directory that we pass to Tradefed's
`--tests-dir` flag. This directory primarily contains all files
installed for a given test under testcases/.

This approach narrows TF's search scope to better match CI behavior.
It is also a more flexible approach as it allows us to craft the
layout to match what TF expects.

Note that we also include a few additional files in the created test
directory that are installed but _outside_ the per-test directory.
For example `art-run-test-500-instanceof` declares a dependency on
`art-run-test-checker` via the `host_required` attribute. The latter
file is installed in testcases/ but not under the
`art-run-test-500-instanceof` subdirectory. This is not an issue for
Atest standard mode since it searches the TF test dir is rooted at
testcases/.

We address this issue by symlinking all transitive test dependencies
that have a non-empty `suites` attribute into the created directory.
This allows these auxiliary files to be found without putting the
entire testcases/ directory in a Bazel test's runfiles. Future work
could improve the situation by ensuring that the build system creates
self-sufficient per-test directories that contain these files.

This change only impacts the TF test dir but not the runfiles and
environment that is set up prior to launching TF.

Test: m atest && atest-dev art-run-test-004-checker-UnsafeTest18 \
          --experimental-device-driven-test
Test: m atest && atest-dev -c -m --host \
          --host-unit-test-only \
          --bazel-arg=--test_timeout=3000
Bug: 222437352
Change-Id: Id2a6d61edb987e087a4c5513d0b1449fb1e55eec