Add experimental Bazel mode support for Robolectric tests
This change adds an experimental Atest flag and supporting logic to run
Robolectric tests in Bazel mode. This feature is disabled by default and
requires passing the `--experimental-robolectric-test` flag on the Atest
command line.

Note that only 'modern' Robolectric tests that are configured to run
with Tradefed are supported. Legacy Robolectric tests that are run by
Make are not supported.

Test: atest --host atest_unittests
Test: atest -c -m --experimental-robolectric-test \
      BackupFrameworksServicesRoboTests
Bug: 228799625
Change-Id: Ic5157ba8955b506e8fa0941644a198af6bed98e2