Only import modules from the root package
Currently, in soong, all the python files are assembled
into a zip file, and then the root of that zip file is
added to the PYTHONPATH when it's run. However, all the
subfolders of that root directory are also added to the
PYTHONPATH, which is why you could say stuff like
"import atest_metrics" instead of
"from atest import atest_metrics". We want to remove
these erroneous extra directories in the path, so update
atest to not rely on this functionality.

This cl also removes the ability to run atest-src,
because it won't have the proper PYTHONPATH set up.

Bug: 247578564
Bug: 245583294
Test: m atest && atest-dev atest_unittests asuite_metrics_lib_tests asuite_cc_lib_tests aidegen_unittests plugin_lib_unittests --host --no-bazel-mode
Change-Id: I9b23a0ce645c3e483d91de43e94169cd4a433294