Add a tool to convert JaCoCo files to LCOV format
This tool converts JaCoCo exec and class files to LCOV format which is
more suitable for generating coverage reports using `genhtml`.

This tool also handles a JaCoCo limitation when processing classes that
are included in multiple build modules. For example, a class X that's
included in both product and test code. JaCoCo will reject and throw
errors when the class is loaded more than once.

Converting results to LCOV format for each build module and then merging
the results allows us to generate a single report for all Java code.

Note that this tool will only generate coverage for Java and Kotlin
files in any of the provided source paths.

Usage:

jacoco_to_lcov_converter \
    -o <destination> \
    [<execfiles>] \
    [-classfiles <classfile>] \
    [-sourcepath <sourcepath>]

Bug: 254109621
Bug: 255586853
Change-Id: I1cd5c96004400bc47cad14f8b2615a488acd62d2