Use Atest to generate workspace and build dependencies
Change the workspace generation process to rely on Atest to both
generate the workspace and build all dependencies via the -b flag.

This reduces the need for extra bazel calls in this script and reduces
the risk of the query logic breaking or diverging. It also makes
behavior between local and CI testing consistent.

Test: gen_workspace_archive_test.sh
Bug: 236743344
Change-Id: Ibe8be8b30b60f33d84f663742c0e5d928e4f1365