Stop installing target app for instrumentation test
To avoid encountering `invalid method` error while running
instrumentation test, Atest tries to find out and install the target app
to ensure the APIs are consistent between the test and the app.

However, `INSTALL_FAILED_VERSION_DOWNGRADE` could happen when the
version of the installed app is lower than the pre-installed one. This
change is to shut off the feature.

Bug: 258081524

Test: m atest && atest-dev CellBroadcastReceiverUnitTests -c
Change-Id: I3442baabcc3b7125d25e2b0de7e0c12dba470325