Fix errors when generating the Bazel workspace
Atest crashes while generating the Bazel workspace due to module-info entries that have a name that is `None`. This is caused by aosp/2293302 which merged synthetic entries into module-info.

More specifically, this caused `_add_missing_variant_modules` to add an entry for these synthetic modules using the declared name attribute which is absent. This resulted in module-info having an entry with a `None` key.

Besides the incorrect data, these synthetic modules cannot be built with Soong since no such target exists.

This change ensures that module-info does not contain a bad entry by making sure that we default to the 'declared' module name when adding these synthetic variants. We also add checks to make sure that modules without a name attribute are not considered 'real' modules and never get accidentally built.

Bug: 259134329
Test: m atest && atest-dev atest_unittests
      atest-dev --host-unit-test-only .

Change-Id: Iaa50939370c8c749bf7a3d511fdc63e21873dc91