Include runtime dependencies of data binaries
Some data binaries have their own runtime dependencies such as shared
libraries and jars. Not including these in the runtime environment
causes tests to fail since these binaries cannot be executed.

Note that test failures are not apparent when using Atest Bazel mode
locally but only surface when the Bazel workspace is moved to a
different location. For example, running tests on a different CI host.

This inconsistency happens because binaries have `RUNPATH` encoded in
the executable file which references shared libraries using a relative
path. When run locally the shared objects can be found unlike when run
on a different host.

Test: /google/data/ro/projects/android/fetch_artifact
  --bid P45595658
  --target aosp_cf_x86_64_bazel_workspace
  'atest_bazel_workspace.tar.gz' &&
  tar -xf atest_bazel_workspace.tar.gz &&
  cd out/atest_bazel_workspace &&
  ./bazel.sh test
  //frameworks/native/libs/binder/tests:binderRpcTestNoKernel_host
  --test_arg=--test-arg=com.android.tradefed.testtype.HostGTest:native-test-timeout:60m
Test: m atest && atest-dev -c -m --host --host-unit-test-only
Fix: 260937596
Change-Id: Id764dd12c36a99155e5b4dd23cd58997ac96bb15