Stop building unnecessary modules in Bazel mode
Atest adds a static list of build-time dependencies that may be
unnecessary for running a test.

This change adapts Bazel mode's behavior to no longer include these
redundant modules since we can tell which runtime dependencies a test
actually needs.

Reducing the build set reduces invocation times. For example,
CtsNNAPITestCases build time drops from 852.2s to 706.6s (-17%).

The test runner interface is modified to accept the tests that it will
run. This makes it possible to push the logic that appends the static
build targets into the runners that require it. This subtly changes
the semantics of the `get_test_runner_build_reqs` to not only return
the runner's own build dependencies but those required for the test
modules themselves.

Test: m atest && atest-dev CtsNNAPITestCases \
  --bazel-arg=--test_timeout=3000 --host
Test: Test: m atest && atest-dev CtsNNAPITestCases \
  --no-bazel-mode --host
Test: m atest && atest-dev --host --host-unit-test-only
Bug: 266957460
Bug: 231694832
Change-Id: Icb6e8ec5d0168963bc2c62806b70e0c0864abdd6