Use supported configs from metadata for modules without install paths
Atest Bazel mode uses the module install paths to deduce the supported
configurations instead of relying on the metadata generated by the
build system. This is a temporary workaround since some test modules,
such as Robolectric tests, export incorrect variants in module-info.

Prior to this change, a module that doesn't have installed paths
assumes that the supported configurations are the same as those of its
dependencies. This is incorrect since device-only modules may have
multi-config dependencies and would be incorrectly mistaken as a
multi-config module.

For example, `libandroid_log_sys.rlib-std` is a device-only module
without any installed paths that depends on `libstd` that has both
host and device variants. Atest Bazel mode mistakenly marks
`libandroid_log_sys.rlib-std` as a multi-config module when it only
supports the device variant.

Test: m atest && atest-dev --host --host-unit-test-only
Test: m atest && atest-dev --host --host-unit-test-only \
  --experimental-remote
Change-Id: Ibdc427cb796e4e25f0e33338bf21dad6e12cf892