Reduce Bazel workspace generation time
Bazel workspace generation takes time since the `is_testable_module`
function reads config files under the module path to determine if a
module is testable.

This is no longer necessary with aosp/2434852 and aosp/2434932 which
adds the necessary information to module info. And this change can
reduce the time of generating Bazel workspace from 12.5 seconds to
1.7 seconds.

Test: m atest && atest-dev --host atest_unittests
Bug: 268670243
Change-Id: I62aca49837de1ab2ede39453b77257e307dbfbc8