Add metrics for regenerating bazel workspace time
Currently, Bazel workspaces are only generated(or regenerated) when
module-ino or feature flags change. But the metric of
BAZEL_WORKSPACE_GENERATE_TIME collect all of the workspace generation
time even the workspace didn't generate. This is inaccurate and gives us
the impression that the time to generate workspace is very fast.

This change add a metric of FULL_GENERATE_BAZEL_WORKSPACE_TIME to
collect the time when a Bazel workspace was actually generated(or
regenerated).

Test: m atest && atest-dev --host atest_unittests
Bug: 269062306
Change-Id: I721b9297320cea228fdd4d03113fa23e988e4368