Add rules to run tests under a device environment
Device tests require connected devices to interact with. The type of environment can vary from a single locally connected device to a remote execution endpoint that accepts jobs for running in some remote lab.

This change implements the framework for defining these environments and the corresponding rule for running test under them.

Test: m atest && atest-dev HelloWorldTests --experimental-device-driven-test
Bug: 260997858
Change-Id: Id2ba975dca2081c0a0710c1ddaa17db81dc3da36