Fix pylint `method-cache-max-size-none` warning
Using @functools.lru_cache without an explicit max cache size can cause
memory leaks as all the arguments to the method, including self, are
kept alive indefinitely.

This change specifies a `maxsize` of 128 to fix the issue. It increases
the average Bazel workspace generation time from 1.9149s to 1.9230s.
Note that removing the cache increases this time to 2.2582s.

Test: PYLINTRC=pylintrc pylint atest/bazel_mode.py
Bug: 271221896
Change-Id: Icc824ddb97f4256f8c4ae92b77e9d87015ab84ce