Regenerate Bazel workspace on dependency files changes
Create a ResourceManager object to manage workspace dependency files.
All files recorded in ResourceManager will be calculated md5 and
written in checksum file. The checksum file will be checked when the
workspace is generated, and the Bazel workspace will be regenerated
as long as there is a change in the file.

Test: m atest && atest-dev atest_unittests
Bug: 265320036
Change-Id: I3dc9e9f2ca7c095d51865a6d99ba6a1c0adc3c51