Fix Java coverage classfile lookup.
The jacoco-report-classes lookup returns a generator which would be incorrectly be converted into a string when there where multiple items present. This, causes the report generation to have an invalid path on the command line and result in an (invalid) empty lcov data file.

This change converts the generator to a collection and passes the entries to the command line.

Bug: 268247629
Bug: 268246219
Bug: 267170674
Test: m atest && atest-dev CtsNearbyFastPairTestCases
--experimental-coverage
Test: m atest && atest-dev
MediaProviderTests[com.google.android.mediaprovider.apex]
--experimental-coverage
Test: m atest && atest-dev adbd_test[com.google.android.adbd.apex]
--experimental-coverage

Change-Id: I8c1a141a559dac622b0945c60c3fd2f614651c75