atest: Fix parse_test regex
Previously the module name was parsed with a `split(':')` which has been
replace by a regex. The regex only consider the module name with r'\w\/.'
which leads module name with a '-' in the middle to be malformatted such
as pts-bot.

`atest pts-bot:<test_name>` is no longer working since atest is looking
for `bot` instead of `pts-bot`.

Modified the regex accordingly to take everything before the ':'.

Test: TreeHugher
Bug: 276813900

Change-Id: Idaf096bf0e9ab4ba680b7b315580b949563f9691