Add tracing to key runner activities
Use Tradefed's support for generating Perfetto traces in the test
runner. Note that this does not trace the Bazel executable or any
processes forked by it.

This change moves timeout handling to when processes are started to
simplify call sites and add automatic tracing for any created
sub-processes. Any timeouts are now logged instead of throwing an
exception which changes the reported failure condition.

Test: atest --experimental-bes-publish bazel-test-runner-tests
Bug: 276367158
Change-Id: I9b3fd8e6bdecd6ad2ce1646b5cadc6562cbd5f66