Treat empty tests in Bazel mode as passing.
In non-Bazel mode, empty tests are treated as passing. In bazel mode,
this code attempts to treat empty tests as failing but doesn't do it
correctly, resulting in the results reporter becoming confused. Here
are a few of the issues:

  - Bazel thinks the tests are failing but TF thinks they're passing
  - The workunits show up as successful even though Bazel failed
  - No errors are shown in the results reporter.

Treating empty tests as passing resolves all these issues for a wide
number of tests that happen to have zero tests in them, while also
addressing the broken TF reporting.

Bug: b/250631336
Test: Trybot run confirms that many more tests are passing
Test: atest-dev --host bazel-test-runner-tests bazel-result-reporter-tests
Change-Id: I1498b17deb00f30864e1d97d604c1890c7e5bca5