Generate module_info_target correctly
Determining whether the product_out is relative to the build_top was
implemented by `find` method, which is very insecure when the build_top
and the product_out has the same prefix but they are not relative, e.g.
build_top == '/path/to/top'
product_out = '/path/to/top-out'

This change uses pathlib to determine whether product_out is the child
dir of the build_top.

Bug: 276784713
Test: source build/envsetup.sh; lunch
Test: export OUT_DIR=${ANDROID_BUILD_TOP}-out
Test: atest-dev hello_world_test --host -c
Change-Id: I130adbc8174d4227486a1f50a8fb4c32072d72df