Set TF's log output directory from BazelTest
TF generates its output directory structure based on the test-tag and
build-id parameters. It used to be "stub/-1/stub/" but some recent
changes to the atest local min config changed this and broke BazelTest.

In order to make BazelTest more stable we set these parameters
ourselves so that we know what to look for when looking for test outputs and don't have to hardcode or guess.

Test: bazel-test-runner-tests
Bug: 251279690
Change-Id: Ibc2887cc2ea45b4bd08cd881018d516548a12ba6