Fix TF log output path in BazelTest
BazelTest cannot find the log files because there is an extra directory
in their path. This started happening when TF started giving a default
value to the branch option in the BootstrapBuildProvider. Set this flag
ourselves in BazelTest in order for the logs to be in a predictable
location.

Test: atest bazel-test-runner-tests
Bug: 285029491
Change-Id: Iae031d71029e02cd0ac537164df43d8881924dfd
Merged-in: Iae031d71029e02cd0ac537164df43d8881924dfd