Fix bug when running non-parameterized Java tests
All Java tests with a method filter in the pattern are being incorrectly
interpreted as parameterized tests due to a missing regex `match` call.
This causes Atest to add a '*' suffix to the TF filter in an attempt to
run all parameterized tests without requiring the user to specify the
exact test name. TF then uses the `tests_regex` instead of the `class`
instrumentation argument when running the test.

Bug: 285399002
Test: m atest ; atest-dev atest_unittests --host
Test: m atest ; atest-dev HealthFitnessUnitTests:\
com.android.server.healthconnect.HealthConnectManagerServiceTest#testUserSwitch (Should not pass * to atest-filter)
Test: m atest ; atest-dev -c CtsWidgetTestCases:\
      android.widget.cts.TextViewPrecomputedTextTest#setText (Should
      pass * to atest-filter)

Change-Id: Ie25937b647441f46d8d469039f1c55120e969b69