Fix incorrect Tradefed invocation attributes
Bazel test executables each run a TF subprocess which writes out its invocation data in proto file. This data is then being merged with the data from other similar invocations along with their properties.

For example, the merged invocation properties will contain a list of timestamp pairs with one of these selected for the top-level invocation. The selected pair is the execution time for a single test and does not correctly represent the invocation's total time.

This change configures TF to not add any per-target invocation data to the top-level context.

Bug: 284294864
Test: Testing in CI
Change-Id: I2d48f5ada12d2a9aedb2730827acf3f91342964d