Stop downloading artifacts during build time
Bazel is downloading Python and Java rules from online repositories
during the build. This causes spurious failures on build hosts that do
not have access to them. It also goes against policies to download
artifacts during the build.

This change configures the workspace with the Roboleaf workspace python
stub and the existing java rules already in AOSP to disable downloading.

Test: atest-dev atest_unittests and presubmit
Bug: 289538639
Change-Id: I41663ae3dc3895aa69422143e6fe0cb62db2e32f