Easily access test logs in Bazel mode like standard mode.
Previously, Atest was unable to display all logs correctly because the
test logs in Bazel mode and Standard mode were stored in different
locations. This change moves the test logs generated by Bazel mode to
the same location as Standard mode, making the same convenience to
access and search the logs in both modes.

Bug: 279869944
Bug: 278758359
Test: m atest && atest-dev atest_unittests asuite_cc_lib_tests --host
Change-Id: I3763c770297cc458ac64baf472b0cef36c15d3fe