Skeleton code for adevice.
This is a little more than skeleton code and isn't as clean
as the bare essentials, but I like to have at least one feature
with a new directory.

This does a minimal of the data structures used to represent
fingerprints of directory trees.
The algorithm could be better, so don't focus on that.
I was just playing with rust data structures.

FYI help output:

% ANDROID_PRODUCT_OUT= ~/main/out/host/linux-x86/bin/adevice --help
Tool to push your rebuilt modules to your device.

Usage: adevice [OPTIONS] [PRODUCT_OUT] [DUMP_FILE_PREFIX] <COMMAND>

Commands:
  status         Shows the file differences between build tree and host
  show-actions   Show the actions that would be run
  update-device  Update the device
  help           Print this message or the help of the given subcommand(s)

Arguments:
  [PRODUCT_OUT]       If unset defaults to ANDROID_PRODUCT_OUT env variable
  [DUMP_FILE_PREFIX]

Options:
  -p, --partitions <PARTITIONS>  Partitions in the product tree to sync. Repeat arg or comma-separate [default: system]
  -h, --help                     Print help

Test: `atest`  # runs both tests because of test_mapping
Test: `m adevice; ANDROID_PRODUCT_OUT= ~/main/out/host/linux-x86/bin/adevice --help ; echo "$?" `
Test: `m adevice; ~/main/out/host/linux-x86/bin/adevice status`

Change-Id: I5684c06c020fef310e0b0922d0eb6e15076b838b