Suppress Unsupported Runner Message for BazelTestRunner and RoboleafTestRunner
This change enhances the output behavior for BazelTestRunner and
RoboleafTestRunner by suppressing the generic "Pretty output does not
support {runner_name}." message when encountering these runners. This
enhancement improves the clarity and user experience of the output,
especially for test runners that are Bazel-based.

Test: m atest; atest-dev atest_unittests
Test: m atest; atest-dev HelloWorldHostTest
Bug: 229300085
Bug: 295973348

Displays a message indicating that the provided runner name is not supported
    for pretty output. If the runner is BazelTestRunner or RoboleafTestRunner,
    no message is printed.

Change-Id: I88728b0fef0460bf29049d3ffa96ac4e2d1620f8