Added some profiling data to output.
When I use cloud based avds, it takes 2 minutes to do an update.
I wanted to emit profile data so if people complain about the time,
we can be sure where it comes from.

On real device, it is about twice as fast (and fingerprinting gets
really fast).

Sample output below:

Test: atest

 Operation profile: (secs)
    	Device Fingerprint - 0
    	Host fingerprint - 0
    	Ninja - 2
    	Adb Cmds - 11
    	Reboot - 4
    	Restart - 33
    	TOTAL - 53

Change-Id: I22117fe8d70ac2700f8fe5e48db5c1d5484784c8